feat: Move SharedData to its own module
